web/handlers/login.go
package handlers

import (
	"crypto/md5"
	"encoding/hex"
	"net/http"

	"github.com/gin-gonic/gin"
)

type Login struct {
	User     string `json:"username"  binding:"required"`
	Password string `json:"password" binding:"required"`
}

func (h *apiHandler) Login(c *gin.Context) {
	var json Login
	if err := c.ShouldBindJSON(&json); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"user":    json.User,
		"session": getMD5Hash(json.User),
	})
	return
}

func getMD5Hash(text string) string {
	hash := md5.Sum([]byte(text))
	return hex.EncodeToString(hash[:])
}
